Refs #3 -- Add new attr to store orig priority
Add `priority_orig` into `struct thread` to store original priority.
We don't need to know the original priority except when we retrieve or
change priority, but just know the representative priority value either
if that value is donated or original.

Donation is not implemented yet and `thread_set_priority` is modified to
deal with donated priority as well as original priority that is already
implemented.